requirements.txt
# =============================================================================
# AI Financial Analysis Server - Dependencies
# =============================================================================

# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ML & Data
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.44.0
joblib>=1.3.0

# PDF Generation
fpdf2>=2.7.0

# LLM
openai>=1.10.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Utils
python-dotenv>=1.0.0

# Azure (데이터/모델 다운로드용)
azure-storage-blob>=12.19.0

# Logging
loguru>=0.7.0

# Testing (dev)
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Code Quality (dev)
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0
